refactor(ext/http): extract 02_websocket.ts from 01_http.js
